# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ConnectionRead": {
        "properties": {
          "client_id": {
            "title": "Client Id",
            "type": "string"
          },
          "client_name": {
            "title": "Client Name",
            "type": "string"
          },
          "connected_at": {
            "format": "date-time",
            "title": "Connected At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "title": "Scopes",
            "type": "array"
          }
        },
        "required": [
          "id",
          "client_id",
          "client_name",
          "connected_at",
          "scopes"
        ],
        "title": "ConnectionRead",
        "type": "object"
      },
...
      "OAuthFailure": {
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          },
          "error_description": {
            "title": "Error Description",
            "type": "string"
          }
        },
        "required": [
          "error",
          "error_description"
        ],
        "title": "OAuthFailure",
        "type": "object"
      },
      "OAuthMetadata": {
        "properties": {
          "authorization_endpoint": {
            "title": "Authorization Endpoint",
            "type": "string"
          },
          "authorization_response_iss_parameter_supported": {
            "default": true,
            "title": "Authorization Response Iss Parameter Supported",
            "type": "boolean"
          },
          "client_id_metadata_document_supported": {
            "default": true,
            "title": "Client Id Metadata Document Supported",
            "type": "boolean"
          },
          "code_challenge_methods_supported": {
            "default": [
              "S256"
            ],
            "items": {
              "type": "string"
            },
            "title": "Code Challenge Methods Supported",
            "type": "array"
          },
          "grant_types_supported": {
            "default": [
              "authorization_code",
              "refresh_token"
            ],
            "items": {
              "type": "string"
            },
            "title": "Grant Types Supported",
            "type": "array"
          },
          "issuer": {
            "title": "Issuer",
            "type": "string"
          },
          "registration_endpoint": {
            "title": "Registration Endpoint",
            "type": "string"
          },
          "response_types_supported": {
            "default": [
              "code"
            ],
            "items": {
              "type": "string"
            },
            "title": "Response Types Supported",
            "type": "array"
          },
          "revocation_endpoint": {
            "title": "Revocation Endpoint",
            "type": "string"
          },
          "revocation_endpoint_auth_methods_supported": {
            "default": [
              "none"
            ],
            "items": {
              "type": "string"
            },
            "title": "Revocation Endpoint Auth Methods Supported",
            "type": "array"
          },
          "scopes_supported": {
            "default": [
              "shopping_list",
              "offline_access"
            ],
            "items": {
              "type": "string"
            },
            "title": "Scopes Supported",
            "type": "array"
          },
          "token_endpoint": {
            "title": "Token Endpoint",
            "type": "string"
          },
          "token_endpoint_auth_methods_supported": {
            "default": [
              "none"
            ],
            "items": {
              "type": "string"
            },
            "title": "Token Endpoint Auth Methods Supported",
            "type": "array"
          }
        },
        "required": [
          "issuer",
          "authorization_endpoint",
          "token_endpoint",
          "registration_endpoint",
          "revocation_endpoint"
        ],
        "title": "OAuthMetadata",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "in": "cookie",
        "name": "__Host-gl_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "GroceryList",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/.well-known/oauth-authorization-server": {
      "get": {
        "operationId": "metadata__well_known_oauth_authorization_server_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthMetadata"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthFailure"
                }
              }
            },
            "description": "Bad Request"
          }
        },
        "summary": "Metadata",
        "tags": [
          "oauth"
        ]
      }
    },
...
    "/api/oauth/connections": {
      "get": {
        "operationId": "connections_api_oauth_connections_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConnectionRead"
                  },
                  "title": "Response Connections Api Oauth Connections Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthFailure"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Требуется вход или неверные учётные данные"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Объект не найден"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Конфликт"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Некорректный запрос"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Внутренняя ошибка"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Connections",
        "tags": [
          "oauth"
        ]
      }
    },
    "/api/oauth/connections/{id}": {
      "delete": {
        "operationId": "disconnect_api_oauth_connections__id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Device-Id",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "X-Device-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthFailure"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Требуется вход или неверные учётные данные"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Объект не найден"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Конфликт"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Некорректный запрос"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Внутренняя ошибка"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Disconnect",
        "tags": [
          "oauth"
        ]
      }
    },
    "/api/oauth/consent": {
      "post": {
        "operationId": "consent_api_oauth_consent_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Device-Id",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "X-Device-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedirectRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthFailure"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Требуется вход или неверные учётные данные"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Объект не найден"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Конфликт"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Некорректный запрос"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Внутренняя ошибка"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Consent",
        "tags": [
          "oauth"
        ]
      }
    },
    "/api/oauth/requests/{id}": {
      "get": {
        "operationId": "consent_info_api_oauth_requests__id__get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthFailure"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Требуется вход или неверные учётные данные"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Объект не найден"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Конфликт"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Некорректный запрос"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Внутренняя ошибка"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Consent Info",
        "tags": [
          "oauth"
        ]
      }
    },
...
    "/oauth/authorize": {
      "get": {
        "operationId": "authorize_oauth_authorize_get",
        "parameters": [
          {
            "in": "query",
            "name": "client_id",
            "required": true,
            "schema": {
              "maxLength": 2048,
              "minLength": 1,
              "title": "Client Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "redirect_uri",
            "required": true,
            "schema": {
              "maxLength": 2048,
              "minLength": 1,
              "title": "Redirect Uri",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "response_type",
            "required": true,
            "schema": {
              "title": "Response Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "code_challenge",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Code Challenge",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "code_challenge_method",
            "required": true,
            "schema": {
              "title": "Code Challenge Method",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "resource",
            "required": true,
            "schema": {
              "maxLength": 2048,
              "minLength": 1,
              "title": "Resource",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "shopping_list",
              "maxLength": 255,
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2048,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          }
        ],
        "responses": {
          "303": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthFailure"
                }
              }
            },
            "description": "Bad Request"
          }
        },
        "summary": "Authorize",
        "tags": [
          "oauth"
        ]
      }
    },
    "/oauth/register": {
      "post": {
        "operationId": "register_oauth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "client_name": {
                    "maxLength": 255,
                    "minLength": 1,
                    "title": "Client Name",
                    "type": "string"
                  },
                  "grant_types": {
                    "default": [
                      "authorization_code",
                      "refresh_token"
                    ],
                    "items": {
                      "enum": [
                        "authorization_code",
                        "refresh_token"
                      ],
                      "type": "string"
                    },
                    "title": "Grant Types",
                    "type": "array"
                  },
                  "redirect_uris": {
                    "items": {
                      "type": "string"
                    },
                    "maxItems": 20,
                    "minItems": 1,
                    "title": "Redirect Uris",
                    "type": "array"
                  },
                  "response_types": {
                    "default": [
                      "code"
                    ],
                    "items": {
                      "const": "code",
                      "type": "string"
                    },
                    "title": "Response Types",
                    "type": "array"
                  },
                  "token_endpoint_auth_method": {
                    "const": "none",
                    "default": "none",
                    "title": "Token Endpoint Auth Method",
                    "type": "string"
                  }
                },
                "required": [
                  "client_name",
                  "redirect_uris"
                ],
                "title": "ClientMetadata",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientRegistered"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthFailure"
                }
              }
            },
            "description": "Bad Request"
          }
        },
        "summary": "Register",
        "tags": [
          "oauth"
        ]
      }
    },
    "/oauth/revoke": {
      "post": {
        "operationId": "revoke_oauth_revoke_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "properties": {
                  "client_id": {
                    "maxLength": 2048,
                    "minLength": 1,
                    "title": "Client Id",
                    "type": "string"
                  },
                  "token": {
                    "maxLength": 256,
                    "minLength": 1,
                    "title": "Token",
                    "type": "string"
                  },
                  "token_type_hint": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "title": "Token Type Hint"
                  }
                },
                "required": [
                  "client_id",
                  "token"
                ],
                "title": "RevokeRequest",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthFailure"
                }
              }
            },
            "description": "Bad Request"
          }
        },
        "summary": "Revoke",
        "tags": [
          "oauth"
        ]
      }
    },
    "/oauth/token": {
      "post": {
        "operationId": "token_oauth_token_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "properties": {
                  "client_id": {
                    "maxLength": 2048,
                    "minLength": 1,
                    "title": "Client Id",
                    "type": "string"
                  },
                  "code": {
                    "default": "",
                    "maxLength": 256,
                    "title": "Code",
                    "type": "string"
                  },
                  "code_verifier": {
                    "default": "",
                    "maxLength": 128,
                    "title": "Code Verifier",
                    "type": "string"
                  },
                  "grant_type": {
                    "title": "Grant Type",
                    "type": "string"
                  },
                  "redirect_uri": {
                    "default": "",
                    "maxLength": 2048,
                    "title": "Redirect Uri",
                    "type": "string"
                  },
                  "refresh_token": {
                    "default": "",
                    "maxLength": 256,
                    "title": "Refresh Token",
                    "type": "string"
                  },
                  "resource": {
                    "maxLength": 2048,
                    "minLength": 1,
                    "title": "Resource",
                    "type": "string"
                  },
                  "scope": {
                    "anyOf": [
                      {
                        "maxLength": 255,
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "title": "Scope"
                  }
                },
                "required": [
                  "grant_type",
                  "client_id",
                  "resource"
                ],
                "title": "TokenRequest",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthTokenRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthFailure"
                }
              }
            },
            "description": "Bad Request"
          }
        },
        "summary": "Token",
        "tags": [
          "oauth"
        ]
      }
    }
  }
}